SourceCode/requirement.txt
pytube3
win32clipboard